+++ src/core/engine/index.ts
@@ -7,4 +7,5 @@
 export * from "./evaluate";
 export * from "./magic";
 export * from "./recruitment";
+export * from "./slavery";
 export * from "./munitions";